requirements.txt
pillow
flask
flask-restful
keras
tensorflow